requirements.txt
librosa
numpy
pandas
pillow
scikit-learn
openpyxl
pyarrow
matplotlib
rich
python-docx
transformers
torch
torchvision
scipy
python-dotenv
groq
anthropic
openai
fpdf2
